user profile image update by uri
