+++ auction-components/share-components/Thumbnail.js
@@ -3,6 +3,7 @@ import React from "react";
 import utils from "@/core/utils";
 
 const Thumbnail = ({ url, size }) => {
+  console.log(utils.thumbnail(url));
   return (
     <Image
       source={utils.thumbnail(url)}
