+++ screens/SearchScreen.js
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from "react-native";
+import React from "react";
+
+const SearchScreen = () => {
+  return (
+    <View>
+      <Text>SearchScreen</Text>
+    </View>
+  );
+};
+
+export default SearchScreen;
+
+const styles = StyleSheet.create({});
